create upload directories if not present

+++ compress.php
@@ -1,10 +1,9 @@
 <?php
 
-use Services\ImageService;
+use ImageService;
 
 require_once './config.php';
 require_once './ImageService.php';
-
 
 $message = '';
 
@@ -27,12 +26,11 @@ function validateUpload(): array {
 
   if (!in_array($imageFileType, ALLOWED_FORMATS)) throw new Exception("Sorry, only the following formats are allowed:" . implode(', ', ALLOWED_FORMATS), 415);
 
-  $target_dir = "uploads/";
-  $targetPath = $target_dir . $originalName;
+  if (!is_dir(UPLOAD_DIR)) mkdir(UPLOAD_DIR);
+  $targetPath = UPLOAD_DIR . $originalName;
   if (!move_uploaded_file($tmpPath, $targetPath)) throw new Exception("Sorry, there has been an error.", 500);
   return [$targetPath, $originalName];
 }
-
 
 function getFileDimesions($path) {
   $imgsize = getimagesize($path);
@@ -45,28 +43,26 @@ function getFileDimesions($path) {
   $widthAdjustment = min(1, $maxWidth / $width);
   $adjustment = min($widthAdjustment, $heightAdjustment);
 
-
   return [$width * $adjustment, $height * $adjustment];
 }
 
-
 function downloadFile() {
+  if (!is_dir(OUTPUT_DIR)) mkdir(OUTPUT_DIR);
   try {
     [$uncompressedPath, $displayName] = validateUpload();
     [$width, $height] = getFileDimesions($uncompressedPath);
     $format = $_POST['format'] ?? DEFAULT_FORMAT;
-    $compressedPath = ImageService::resize_image($uncompressedPath, $width, $height, $format, './compressed/' . time());
+    $compressedPath = ImageService::resize_image($uncompressedPath, $width, $height, $format, OUTPUT_DIR . '/' . time());
     $fileSize = filesize($compressedPath);
 
     header("Cache-Control: private");
     header("Content-Type: application/stream");
     header("Content-Length: " . $fileSize);
-    header("Content-Disposition: attachment; filename=" . ImageService::replace_extension($displayName,$format));
+    header("Content-Disposition: attachment; filename=" . ImageService::replace_extension($displayName, $format));
 
     readfile($compressedPath);
-    unlink($compressedPath);
+    if (!KEEP_COMPRESSED) unlink($compressedPath);
     if (!KEEP_ORIGINAL) unlink($uncompressedPath);
-
     exit();
   } catch (Exception $e) {
     global $message;
@@ -74,7 +70,6 @@ function downloadFile() {
   }
 }
 downloadFile();
-
 ?>
 
 <!DOCTYPE html>
